Update typing to allow operation inside Python 3.7
