+++ requirements.txt
@@ -11,6 +11,8 @@ jsonschema==3.2.0
 polyswarm-artifact==1.2.1
 psycopg2-binary==2.8.4
 pytest==4.0.0
+pycryptodome==3.8.1
+py-evm==0.3.0a5
 python-consul==1.1.0
 python-json-logger==0.1.9
 PyYaml==4.2b4
