+++ tox.ini
@@ -4,7 +4,7 @@ envlist = py37
 
 [testenv]
 commands = pytest -s
-install_command=python -m pip install --no-use-pep517 {opts} {packages}
+install_command=python -m pip install {opts} {packages}
 usedevelop=True
 
 [isort]
@@ -48,3 +48,4 @@ commands=
   mypy src/polyswarmd --config-file {toxinidir}/mypy.ini
   flake8 {toxinidir}/src/polyswarmd
   isort --recursive --check-only --diff {toxinidir}/src/polyswarmd
+
